Rename react-auth0.js -> Auth0Provider.js
