+++ app/components/Header/UserHeader.js
@@ -11,7 +11,7 @@ import { updateProfileDedication } from '../../actions/updateUserProfile';
 import { updateRoute } from '../../helpers/routerHelper';
 import { currentUserProfileSelector, userFeedsSelector } from '../../selectors';
 import { profileImageUrl } from '../../utils/profileImage';
-import { useAuth0 } from '../auth0/react-auth0';
+import { useAuth0 } from '../auth0/Auth0Provider';
 import Popover from '../Common/Popover';
 import ProfilePickerModal from '../EditUserProfile/dedicate-trees/ProfilePickerModal';
 import Notification from './Notification';
